social_network/requirements.txt
django==5.0.2
djangorestframework==3.14.0
pillow==10.2.0
psycopg2-binary==2.9.9
geopy==2.4.1

